feat: workspace nested routing
