+++ client/src/page/Workspace/template/index.ts
@@ -1,0 +1,4 @@
+import Channel from './Channel'
+import ChannelBrowser from './ChannelBrowser'
+
+export { Channel, ChannelBrowser }
