+++ client/src/page/WorkspaceBrowser/WorkspaceJoinPage.tsx
